refactor(AHWR-441): Minor improvements ahead of retry of ESM refactor
Remove unused permissions file
Remove endemics enabled flag
Add ignore assets tag on logging
Move header plugin options into plugin file
Remove blipp again
