--- app/auth/permissions.js
@@ -1,9 +1,0 @@
-const permissions = {
-  administrator: 'administrator',
-  processor: 'processor',
-  user: 'user',
-  recommender: 'recommender',
-  authoriser: 'authoriser'
-}
-
-module.exports = permissions
