+++ app/plugins/header.js
@@ -1,3 +1,14 @@
+const config = require('../config')
+
+const getSecurityPolicy = () => "default-src 'self';" +
+  "object-src 'none';" +
+  "script-src 'self' www.google-analytics.com *.googletagmanager.com ajax.googleapis.com *.googletagmanager.com/gtm.js 'unsafe-inline' 'unsafe-eval' 'unsafe-hashes';" +
+  "form-action 'self';" +
+  "base-uri 'self';" +
+  "connect-src 'self' *.google-analytics.com *.analytics.google.com *.googletagmanager.com" +
+  "style-src 'self' 'unsafe-inline' tagmanager.google.com *.googleapis.com;" +
+  "img-src 'self' *.google-analytics.com *.googletagmanager.com;"
+
 module.exports = {
   plugin: {
     name: 'header',
@@ -10,5 +21,24 @@ module.exports = {
         return h.continue
       })
     }
+  },
+  options: {
+    keys: [
+      { key: 'X-Frame-Options', value: 'deny' },
+      { key: 'X-Content-Type-Options', value: 'nosniff' },
+      { key: 'Access-Control-Allow-Origin', value: config.serviceUri },
+      { key: 'Cross-Origin-Opener-Policy', value: 'same-origin' },
+      { key: 'Cross-Origin-Embedder-Policy', value: 'require-corp' },
+      { key: 'X-Robots-Tag', value: 'noindex, nofollow' },
+      { key: 'X-XSS-Protection', value: '1; mode=block' },
+      { key: 'Strict-Transport-Security', value: 'max-age=31536000;' },
+      { key: 'Cache-Control', value: 'no-cache' },
+      { key: 'Referrer-Policy', value: 'no-referrer' },
+      { key: 'Permissions-Policy', value: 'Interest-Cohort=()' },
+      {
+        key: 'Content-Security-Policy',
+        value: getSecurityPolicy()
+      }
+    ]
   }
 }
