+++ app/plugins/router.js
@@ -1,22 +1,15 @@
-const config = require('../config')
-
-let routes = [].concat(
+const routes = [].concat(
   require('../routes/healthy'),
-  require('../routes/healthz')
+  require('../routes/healthz'),
+  require('../routes/assets'),
+  require('../routes/cookies'),
+  require('../routes/index'),
+  require('../routes/check-details'),
+  require('../routes/update-details'),
+  require('../routes/signin-oidc'),
+  require('../routes/vet-visits'),
+  require('../routes/download-application')
 )
-
-if (config.endemics.enabled) {
-  routes = routes.concat(
-    require('../routes/assets'),
-    require('../routes/cookies'),
-    require('../routes/index'),
-    require('../routes/check-details'),
-    require('../routes/update-details'),
-    require('../routes/signin-oidc'),
-    require('../routes/vet-visits'),
-    require('../routes/download-application')
-  )
-}
 
 module.exports = {
   plugin: {
