+++ app/routes/assets.js
@@ -3,6 +3,7 @@ module.exports = {
   path: '/assets/{path*}',
   options: {
     auth: false,
+    tags: ['assets'],
     handler: {
       directory: {
         path: ['app/frontend/dist', 'node_modules/govuk-frontend/dist/govuk/assets']
