+++ docker-compose.override.yaml
@@ -8,7 +8,9 @@ services:
     build:
       target: development
     image: ffc-ahwr-dashboard-development
-    container_name: ffc-ahwr-dashboard-development   
+    container_name: ffc-ahwr-dashboard-development
+    environment:
+      USE_PRETTY_PRINT: ${USE_PRETTY_PRINT:-true}
     ports:
       - "3003:3000"
       - "9232:9229"
