+++ helm/ffc-ahwr-dashboard/values.yaml
@@ -70,7 +70,6 @@ container:
   apimClientId:
   apimClientSecret:
   apimScope:
-  endemicsEnabled:
   azureStorageConnectionString:
   azureStorageAccountName:
   azureStorageApplicationDocumentsContainerName:
