+++ test/unit/app/plugins/router.test.js
@@ -11,18 +11,6 @@ describe('routes plugin test', () => {
   })
 
   test('routes included', async () => {
-    const createServer = require('../../../../app/server')
-    const server = await createServer()
-    const routePaths = []
-    server.table().forEach((element) => {
-      routePaths.push(element.path)
-    })
-    expect(routePaths).toEqual([
-      '/healthy', '/healthz'
-    ])
-  })
-
-  test('routes included - endemics enabled', async () => {
     jest.mock('../../../../app/config', () => ({
       ...jest.requireActual('../../../../app/config'),
       endemics: {
